refactor: reach 전략 패키지 경로 단순화
