+++ src/main/java/model/piece/Guard.java
@@ -1,10 +1,9 @@
 package model.piece;
 
 import model.Team;
-import model.coordinate.PalacePositions;
 import model.coordinate.Position;
-import model.piece.strategy.reach.GuardKingReach;
-import model.piece.strategy.reach.ReachStrategy;
+import model.piece.strategy.GuardKingReach;
+import model.piece.strategy.ReachStrategy;
 
 import java.util.List;
