+++ src/main/java/model/piece/strategy/AnimalReach.java
@@ -1,4 +1,4 @@
-package model.piece.strategy.reach;
+package model.piece.strategy;
 
 // 상(2), 마(1): 직선 1칸 + 대각선 N칸
 public class AnimalReach implements ReachStrategy {
