+++ src/main/java/model/piece/strategy/LinearReach.java
@@ -1,4 +1,4 @@
-package model.piece.strategy.reach;
+package model.piece.strategy;
 
 // 차, 포: 상하좌우 N칸 직선
 public class LinearReach implements ReachStrategy {
