+++ src/main/java/model/piece/strategy/ReachStrategy.java
@@ -1,4 +1,4 @@
-package model.piece.strategy.reach;
+package model.piece.strategy;
 
 public interface ReachStrategy {
     boolean isReachable(int rowDiff, int colDiff);
